Archive/requirements.txt
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0
albumentations==1.4.8
stringzilla==4.4.0
opencv-python==4.10.0.84
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.2
scipy==1.16.3
cachetools==5.5.0
torchmetrics==1.3.2
pyyaml==6.0.2
tqdm==4.66.5
rich==13.7.1
loguru==0.7.2
matplotlib==3.9.2
seaborn==0.13.2
pillow==10.4.0
fastapi==0.114.1
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
streamlit==1.38.0
onnx==1.19.1
onnxruntime==1.23.2
black==24.8.0
ruff==0.6.9
openpyxl==3.1.5
pytest==8.3.3
